Take BC.1 out of labels again
The lineage BC.1 has been withdrawn, but maybe only temporarily. I remove it from the labeling to avoid confusion, but it's still mapped in collapsed graphics, just in case it will be re-introduced later.

Also added missing period after "BA.3", so that e.g. "BA.32" would not be counted as subvariant of "BA.3".